Refactor components folder from layout
