+++ src/auth/LoginPage.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import {Link} from 'react-router-dom';
-import Title from '../layout/Title';
-import Subtitle from '../layout/Subtitle';
+import Title from '../components/Title';
+import Subtitle from '../components/Subtitle';
 import LoginForm from './form/LoginForm';
 
 // import {AuthorizedRoute, AuthorizedLink} from './routes/Router';
