--- src/routes/PatrimonyRoutes.js
@@ -1,6 +1,0 @@
-const patrimonyRoutes = {
-  patrimony: '/api/patrimony/',
-  active: '/api/patrimony/active',
-};
-
-export default patrimonyRoutes;
